// models/RestNode.ts
export interface RestNode {
    /**
     * 
     * @type {Array<ActivityObject>}
     * @memberof RestNode
     */
    activities?: Array<ActivityObject>;
    /**
     * 
     * @type {RestLockInfo}
     * @memberof RestNode
     */
    contentLock?: RestLockInfo;
    /**
     * 
     * @type {string}
     * @memberof RestNode
     */
    contentType?: string;
    /**
     * 
     * @type {string}
     * @memberof RestNode
     */
    contentsHash?: string;
    /**
     * 
     * @type {RestContextWorkspace}
     * @memberof RestNode
     */
    contextWorkspace?: RestContextWorkspace;
    /**
     * 
     * @type {RestDataSourceFeatures}
     * @memberof RestNode
     */
    dataSourceFeatures?: RestDataSourceFeatures;
    /**
     * 
     * @type {Array<RestCountMeta>}
     * @memberof RestNode
     */
    folderMeta?: Array<RestCountMeta>;
    /**
     * 
     * @type {string}
     * @memberof RestNode
     */
    hashingMethod?: string;
    /**
     * 
     * @type {RestImageMeta}
     * @memberof RestNode
     */
    imageMeta?: RestImageMeta;
    /**
     * 
     * @type {boolean}
     * @memberof RestNode
     */
    isBookmarked?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RestNode
     */
    isRecycleBin?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RestNode
     */
    isRecycled?: boolean;
    /**
     * 
     * @type {Array<RestJsonMeta>}
     * @memberof RestNode
     */
    metadata?: Array<RestJsonMeta>;
    /**
     * 
     * @type {RestMode}
     * @memberof RestNode
     */
    mode?: RestMode;
    /**
     * 
     * @type {string}
     * @memberof RestNode
     */
    modified?: string;
    /**
     * 
     * @type {string}
     * @memberof RestNode
     */
    path?: string;
    /**
     * 
     * @type {Array<RestFilePreview>}
     * @memberof RestNode
     */
    previews?: Array<RestFilePreview>;
    /**
     * 
     * @type {Array<RestShareLink>}
     * @memberof RestNode
     */
    shares?: Array<RestShareLink>;
    /**
     * 
     * @type {string}
     * @memberof RestNode
     */
    size?: string;
    /**
     * 
     * @type {string}
     * @memberof RestNode
     */
    storageETag?: string;
    /**
     * 
     * @type {Array<ActivitySubscription>}
     * @memberof RestNode
     */
    subscriptions?: Array<ActivitySubscription>;
    /**
     * 
     * @type {TreeNodeType}
     * @memberof RestNode
     */
    type?: TreeNodeType;
    /**
     * 
     * @type {Array<RestUserMeta>}
     * @memberof RestNode
     */
    userMetadata?: Array<RestUserMeta>;
    /**
     * 
     * @type {string}
     * @memberof RestNode
     */
    uuid?: string;
}